working commit

+++ py_payment_api/app/services/__init__.py
@@ -1,3 +1,4 @@
+
 from .services import (
     all_payment_service,
     apple_subscription_service,
@@ -22,13 +23,9 @@ from .store.store_service import StoreService
 from .payment_service import PaymentService
 from .store.items import AvailableItem
 from .payment_service import PaymentData
+from .enums import PaymentProvider
 
-class PaymentProvider(StrEnum):
-    APPLE = "apple"
-    GOOGLE = "google"
-    PAYPAL = "paypal"
-    COINBASE = "coinbase"
-    COINSUB = "coinsub"
+
 
 ITEM_CATEGORY = TypeVar('ITEM_CATEGORY', bound=StrEnum)
 class Services(Generic[ITEM_CATEGORY]):
@@ -38,13 +35,17 @@ class Services(Generic[ITEM_CATEGORY]):
     __google_subscription_service: GoogleSubscriptionService
     __paypal_subscription_service: PaypalSubscriptionService
     __coinsub_subscription_service: CoinsubSubscriptionService
-    def __init__(self):
+    def __init__(self, enabled_providers: list[PaymentProvider]):
         self.__store_service = construct_store_service[ITEM_CATEGORY]()
         self.__payment_service = PaymentService()
-        self.__apple_subscription_service = AppleSubscriptionService()
-        self.__google_subscription_service = GoogleSubscriptionService()
-        self.__paypal_subscription_service = PaypalSubscriptionService()
-        self.__coinsub_subscription_service = CoinsubSubscriptionService()
+        if PaymentProvider.APPLE in enabled_providers:
+            self.__apple_subscription_service = AppleSubscriptionService()
+        if PaymentProvider.GOOGLE in enabled_providers:
+            self.__google_subscription_service = GoogleSubscriptionService()
+        if PaymentProvider.PAYPAL in enabled_providers:
+            self.__paypal_subscription_service = PaypalSubscriptionService()
+        if PaymentProvider.COINSUB in enabled_providers:
+            self.__coinsub_subscription_service = CoinsubSubscriptionService()
 
 
     def  get_items(self, item_category: ITEM_CATEGORY) -> list[AvailableItem]:
@@ -54,67 +55,103 @@ class Services(Generic[ITEM_CATEGORY]):
         return self.__store_service.get_all_items()
     
     def on_apple_payment(self, payment_data: AppleData):
-        self.__payment_service.on_apple_payment(payment_data)
+        if hasattr(self, '__apple_subscription_service'):
+            self.__payment_service.on_apple_payment(payment_data)
+        else:
+            raise ValueError("Apple payment service is not enabled.")
 
     def on_google_payment(self, payment_data: GoogleData):
-        self.__payment_service.on_google_payment(payment_data)
+        if hasattr(self, '__google_subscription_service'):
+            self.__payment_service.on_google_payment(payment_data)
+        else:
+            raise ValueError("Google payment service is not enabled.")
 
     def on_paypal_payment(self, payment_data: PaypalData):
-        self.__payment_service.on_paypal_payment(payment_data)
+        if hasattr(self, '__paypal_subscription_service'):
+            self.__payment_service.on_paypal_payment(payment_data)
+        else:
+            raise ValueError("PayPal payment service is not enabled.")
 
     def on_coinbase_payment(self, payment_data: CoinbaseData):
         self.__payment_service.on_coinbase_payment(payment_data)
 
     def on_coinsub_payment(self, payment_data: CoinSubData):
-        self.__payment_service.on_coinsub_payment(payment_data)
+        if hasattr(self, '__coinsub_subscription_service'):
+            self.__payment_service.on_coinsub_payment(payment_data)
+        else:
+            raise ValueError("CoinSub payment service is not enabled.")
 
     def add_apple_subscription(self, subscription_data: AppleSubscriptionData):
-        self.__apple_subscription_service.add_subscription(subscription_data)
+        if hasattr(self, '__apple_subscription_service'):
+            self.__apple_subscription_service.add_subscription(subscription_data)
+        else:
+            raise ValueError("Apple subscription service is not enabled.")
 
     def add_google_subscription(self, subscription_data: GoogleSubscriptionData):
-        self.__google_subscription_service.add_subscription(subscription_data)
+        if hasattr(self, '__google_subscription_service'):
+            self.__google_subscription_service.add_subscription(subscription_data)
+        else:
+            raise ValueError("Google subscription service is not enabled.")
 
     def add_paypal_subscription(self, subscription_data: PaypalSubscriptionData):
-        self.__paypal_subscription_service.add_subscription(subscription_data)
+        if hasattr(self, '__paypal_subscription_service'):
+            self.__paypal_subscription_service.add_subscription(subscription_data)
+        else:
+            raise ValueError("PayPal subscription service is not enabled.")
 
     def add_coinsub_subscription(self, subscription_data: CoinSubScriptionData):
-        self.__coinsub_subscription_service.add_subscription(subscription_data)
+        if hasattr(self, '__coinsub_subscription_service'):
+            self.__coinsub_subscription_service.add_subscription(subscription_data)
+        else:
+            raise ValueError("CoinSub subscription service is not enabled.")
 
     def remove_apple_subscription(self, user_id: str):
-        self.__apple_subscription_service.remove_subscription(user_id)
+        if hasattr(self, '__apple_subscription_service'):
+            self.__apple_subscription_service.remove_subscription(user_id)
 
     def remove_google_subscription(self, user_id: str):
-        self.__google_subscription_service.remove_subscription(user_id)
+        if hasattr(self, '__google_subscription_service'):
+            self.__google_subscription_service.remove_subscription(user_id)
 
     def remove_paypal_subscription(self, user_id: str):
-        self.__paypal_subscription_service.remove_subscription(user_id)
+        if hasattr(self, '__paypal_subscription_service'):
+            self.__paypal_subscription_service.remove_subscription(user_id)
 
     def remove_coinsub_subscription(self, user_id: str):
-        self.__coinsub_subscription_service.remove_subscription(user_id)
+        if hasattr(self, '__coinsub_subscription_service'):
+            self.__coinsub_subscription_service.remove_subscription(user_id)
 
     def get_apple_subscriptions(self, user_id: str) -> List[AppleSubscriptionData]:
-        return self.__apple_subscription_service.get_subscriptions(user_id)
+        if hasattr(self, '__apple_subscription_service'):
+            return self.__apple_subscription_service.get_subscriptions(user_id)
 
     def get_google_subscriptions(self, user_id: str) -> List[GoogleSubscriptionData]:
-        return self.__google_subscription_service.get_subscriptions(user_id)
+        if hasattr(self, '__google_subscription_service'):
+            return self.__google_subscription_service.get_subscriptions(user_id)
 
     def get_paypal_subscriptions(self, user_id: str) -> List[PaypalSubscriptionData]:
-        return self.__paypal_subscription_service.get_subscriptions(user_id)
+        if hasattr(self, '__paypal_subscription_service'):
+            return self.__paypal_subscription_service.get_subscriptions(user_id)
 
     def get_coinsub_subscriptions(self, user_id: str) -> List[CoinSubScriptionData]:
-        return self.__coinsub_subscription_service.get_subscriptions(user_id)
+        if hasattr(self, '__coinsub_subscription_service'):
+            return self.__coinsub_subscription_service.get_subscriptions(user_id)
     
     def register_apple_subscription_handler(self, handler: AppleSubscriptionHandler):
-        self.__apple_subscription_service.register_handler(handler)
+        if hasattr(self, '__apple_subscription_service'):
+            self.__apple_subscription_service.register_handler(handler)
 
     def register_google_subscription_handler(self, handler: GoogleSubscriptionHandler):
-        self.__google_subscription_service.register_handler(handler)
+        if hasattr(self, '__google_subscription_service'):
+            self.__google_subscription_service.register_handler(handler)
 
     def register_paypal_subscription_handler(self, handler: PaypalSubscriptionHandler):
-        self.__paypal_subscription_service.register_handler(handler)
+        if hasattr(self, '__paypal_subscription_service'):
+            self.__paypal_subscription_service.register_handler(handler)
 
     def register_coinsub_subscription_handler(self, handler: CoinsubSubscriptionHandler):
-        self.__coinsub_subscription_service.register_handler(handler)
+        if hasattr(self, '__coinsub_subscription_service'):
+            self.__coinsub_subscription_service.register_handler(handler)
